adds multiple block/waves schedulers and multiple bug-fix
- Fixes bug which ignores first record of the file with HADOOP
- Enables using input files larger than 2.5G
- Adds FileLeader cache, dramatically increases its perf
- Speedup metadata sqlite access
- Adds support for logical remote reading
- Fixes data leaking
- Multiple refactor and bugfix
- Adds scheduler-multiwave
- Separates logical-blocks sections

Meta-commit squeezed from multiple commits

+++ src/client/vdfs.hh
@@ -62,7 +62,7 @@ class vdfs {
 
     uint32_t write(long, const char*, uint32_t, uint32_t);
     uint32_t write(long, const char*, uint32_t, uint32_t, uint64_t);
-    uint32_t read(long, char*, uint32_t, uint32_t);
+    uint32_t read(long, char*, uint64_t, uint64_t);
 
     model::metadata get_metadata(long fid, int type);
